+++ src/server/packets/packet_server_update_pos.hpp
@@ -5,14 +5,12 @@
 class PacketServerUpdatePos : public Packet
 {
 public:
-    uint16_t    entityId;
-    int16_t     posX;
-    int16_t     posY;
+    uint16_t entityId;
+    int16_t  posX;
+    int16_t  posY;
 
 public:
-    PacketServerUpdatePos(uint16_t entityId, int16_t posX, int16_t posY) : entityId(entityId), posX(posX), posY(posY)
-    {
-    }
+    PacketServerUpdatePos(uint16_t entityId, int16_t posX, int16_t posY) : entityId(entityId), posX(posX), posY(posY) {}
 
     PacketServerUpdatePos()           = default;
     ~PacketServerUpdatePos() override = default;
@@ -20,8 +18,8 @@ public:
     void readPacket(Buffer *buffer) override
     {
         this->entityId = buffer->readU16();
-        this->posX     = buffer->readU16();
-        this->posY     = buffer->readU16();
+        this->posX     = static_cast<int16_t>(buffer->readU16());
+        this->posY     = static_cast<int16_t>(buffer->readU16());
     };
 
     void writePacket(Buffer *buffer) override
